refactor: 사용하지 않는 기능 제거
