+++ src/main/java/blackjack/domain/participant/player/Player.java
@@ -1,19 +1,13 @@
 package blackjack.domain.participant.player;
 
-import static blackjack.domain.game.WinningResult.LOSE;
-import static blackjack.domain.game.WinningResult.TIE;
-import static blackjack.domain.game.WinningResult.WIN;
 import static blackjack.domain.participant.dealer.Dealer.DEALER_NAME;
 import static blackjack.domain.participant.player.AddCardOrNot.NO;
 
-import blackjack.domain.BetAmount;
 import blackjack.domain.Money;
 import blackjack.domain.deck.Deck;
-import blackjack.domain.game.WinningResult;
 import blackjack.domain.participant.Hand;
 import blackjack.domain.participant.Name;
 import blackjack.domain.participant.Participant;
-import blackjack.domain.participant.Score;
 
 public class Player extends Participant {
 
@@ -25,21 +19,6 @@ public class Player extends Participant {
             throw new IllegalArgumentException("플레이어의 이름은 '딜러'일 수 없습니다.");
         }
         this.betAmount = betAmount;
-    }
-
-    public WinningResult combat(Participant dealer) {
-        Score score = calculateScore();
-        Score otherScore = dealer.calculateScore();
-        if (isBust()) {
-            return LOSE;
-        }
-        if (score.isHigherThan(otherScore) || dealer.isBust()) {
-            return WIN;
-        }
-        if (score.equals(otherScore)) {
-            return TIE;
-        }
-        return LOSE;
     }
 
     public void hitAdditionalCardFrom(Deck deck, CardDecisionStrategy cardDecisionStrategy,
