+++ src/main/java/blackjack/domain/participant/player/Players.java
@@ -6,13 +6,10 @@ import static blackjack.domain.participant.FinishedState.STAY;
 
 import blackjack.domain.Money;
 import blackjack.domain.deck.Deck;
-import blackjack.domain.game.WinningResult;
 import blackjack.domain.participant.ParticipantCardsDto;
 import blackjack.domain.participant.ParticipantResultDto;
 import blackjack.domain.participant.dealer.Dealer;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -59,15 +56,6 @@ public class Players {
                 .map(ParticipantResultDto::from)
                 .collect(Collectors.toUnmodifiableList());
     }
-
-    public Map<Player, WinningResult> calculateWinning(Dealer dealer) {
-        Map<Player, WinningResult> playerToResult = new HashMap<>();
-        for (Player player : players) {
-            playerToResult.put(player, player.combat(dealer));
-        }
-        return Collections.unmodifiableMap(playerToResult);
-    }
-
     public List<ParticipantCardsDto> getPlayerCards() {
         return players.stream()
                 .map(ParticipantCardsDto::from)
