+++ src/test/java/blackjack/domain/BetAmountTest.java
@@ -3,6 +3,7 @@ package blackjack.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.domain.participant.player.BetAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
